requirements.txt
fastapi
uvicorn
matplotlib
pyserial
